added API support for fetch the standards

+++ api/src/main/java/org/ednovo/gooru/infrastructure/persistence/hibernate/taxonomy/TaxonomyRespository.java
@@ -135,6 +135,6 @@ public interface TaxonomyRespository extends BaseRepository{
 	
 	Code findCodeByCodeIds(Integer codeId);
 	
-	List<Code> findCodeStartWith(String codeStartWith, Integer depth);
+	List<Code> findCodeStartWith(String codeStartWith, Short depth);
 	
 }
